added federated user link
moved working link into separete component

fixes: #3900
