+++ apps/admin-ui/src/user/user-credentials/FederatedCredentials.tsx
@@ -1,5 +1,3 @@
-import { useState } from "react";
-import { useTranslation } from "react-i18next";
 import {
   Button,
   PageSection,
@@ -13,15 +11,13 @@ import {
   Thead,
   Tr,
 } from "@patternfly/react-table";
+import { useState } from "react";
+import { useTranslation } from "react-i18next";
 
-import type ComponentRepresentation from "@keycloak/keycloak-admin-client/lib/defs/componentRepresentation";
 import type UserRepresentation from "@keycloak/keycloak-admin-client/lib/defs/userRepresentation";
-import { useAdminClient, useFetch } from "../../context/auth/AdminClient";
 import { KeycloakSpinner } from "../../components/keycloak-spinner/KeycloakSpinner";
-import { useAccess } from "../../context/access/Access";
-import { toUserFederationLdap } from "../../user-federation/routes/UserFederationLdap";
-import { useRealm } from "../../context/realm-context/RealmContext";
-import { Link } from "react-router-dom";
+import { useAdminClient, useFetch } from "../../context/auth/AdminClient";
+import { FederatedUserLink } from "../FederatedUserLink";
 
 type FederatedCredentialsProps = {
   user: UserRepresentation;
@@ -34,32 +30,16 @@ export const FederatedCredentials = ({
 }: FederatedCredentialsProps) => {
   const { t } = useTranslation("users");
   const { adminClient } = useAdminClient();
-  const { realm } = useRealm();
 
   const [credentialTypes, setCredentialTypes] = useState<string[]>();
-  const [component, setComponent] = useState<ComponentRepresentation>();
-  const access = useAccess();
 
   useFetch(
-    () =>
-      Promise.all([
-        adminClient.users.getUserStorageCredentialTypes({ id: user.id! }),
-        access.hasAccess("view-realm")
-          ? adminClient.components.findOne({
-              id: (user.federationLink || user.origin)!,
-            })
-          : adminClient.userStorageProvider.name({
-              id: (user.federationLink || user.origin)!,
-            }),
-      ]),
-    ([credentialTypes, component]) => {
-      setCredentialTypes(credentialTypes);
-      setComponent(component);
-    },
+    () => adminClient.users.getUserStorageCredentialTypes({ id: user.id! }),
+    setCredentialTypes,
     []
   );
 
-  if (!credentialTypes || !component) {
+  if (!credentialTypes) {
     return <KeycloakSpinner />;
   }
 
@@ -80,21 +60,7 @@ export const FederatedCredentials = ({
                 <b>{credential}</b>
               </Td>
               <Td>
-                <Button
-                  variant="link"
-                  isDisabled={!access.hasAccess("view-realm")}
-                  component={(props) => (
-                    <Link
-                      {...props}
-                      to={toUserFederationLdap({
-                        id: component.id!,
-                        realm,
-                      })}
-                    />
-                  )}
-                >
-                  {component.name}
-                </Button>
+                <FederatedUserLink user={user} />
               </Td>
               {credential === "password" && (
                 <Td modifier="fitContent">
